the migration structure has been changed, the refresh_tokens table has been added, the phone_number field has been added to the users table, the migrator folder has been moved to pkg and auto-migrations removed from initDB
